remove memory leak...
DDCompactView is anyway pretty much a singleton.... so I changed the way the
DDCompactView(DDLogicalPart) constructor worked so that we don't need to worry
in XMLIdealGeometryESSource about COPYING the whole graph.  I also decided
that we should attempt to use the DDCompactView::clear() method and if it isn't
correct to work on making in so.